Cleanup: Move _Parser to parser.py
